Add support for custom bookmark persistence. (#5)
What
===
Add support for custom bookmark persistence. The existing logic for storing bookmarks in `NSUserDefaults` is moved into a default delegate, and that delegate can be set and replaced with other logic.

Why
===
So that the library doesn't lock the developer into a specific persistence mechanism. The developer can now define a custom delegate and could provide a delegate for storing the bookmarks in a Core Data database, to file, or to network.